add pipe and exceptions handling

+++ src/book/books.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { BadRequestException, Injectable, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { CreateBookDto } from "./dtos/create-book.dto";
@@ -13,19 +13,44 @@ export class BooksService {
         return await this.booksRepository.find();
     }
 
-    async createBook(createBookDto: CreateBookDto): Promise<void> {
-        await this.booksRepository.save(createBookDto);
+    async createBook(createBookDto: CreateBookDto): Promise<Book> {
+        const { title } = createBookDto;
+        const found_book = await this.booksRepository.findOne({ title }); 
+
+        if(found_book) {
+            throw new BadRequestException(`${title} already exists!`);
+        }
+        
+        return await this.booksRepository.save(createBookDto);
     }
 
     async getBook(id: string): Promise<Book> {
+        const found_book = await this.booksRepository.findOne({ id }); 
+
+        if(!found_book) {
+            throw new NotFoundException(`Book with id ${id} doesn't exists.`);
+        }
+
         return await this.booksRepository.findOne(id);
     }
 
     async updateBook(id: string, updateBookDto: UpdateBookDto): Promise<void> {
+        const found_book = await this.booksRepository.findOne({ id }); 
+
+        if(!found_book) {
+            throw new NotFoundException(`Book with id ${id} doesn't exists.`);
+        }
+        
         await this.booksRepository.update(id, updateBookDto);
     }
 
-    async deleteBook(id: string): Promise<void> {
-        await this.booksRepository.delete(id);
-      }
+    async deleteBook(id: string): Promise<any> {
+        const found_book = await this.booksRepository.findOne({ id }); 
+
+        if(!found_book) {
+            throw new NotFoundException(`Book with id ${id} doesn't exists.`);
+        }
+
+        return await this.booksRepository.delete(id);
+    }
 }
